refactor: move Registry to a subpackage
