+++ lib/src/test/java/hu/bme/mit/ftsrg/hypernate/IntegrationTest.java
@@ -7,6 +7,7 @@ import static org.mockito.BDDMockito.then;
 import hu.bme.mit.ftsrg.hypernate.entity.Entity;
 import hu.bme.mit.ftsrg.hypernate.entity.EntityExistsException;
 import hu.bme.mit.ftsrg.hypernate.entity.SerializationException;
+import hu.bme.mit.ftsrg.hypernate.registry.Registry;
 import lombok.Data;
 import org.hyperledger.fabric.shim.ChaincodeStub;
 import org.hyperledger.fabric.shim.ledger.CompositeKey;
